requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.25.2
openai==1.3.7
anthropic==0.7.7
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
markdown==3.5.1